redesign memory structure

+++ main/main.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 // main program for benchmarking
 int main (int argc, char *argv[]) {
-  std::function<int(string)> keyConverter = [](string s){ return s.size(); };
+  std::function<int(string)> keyConverter = [](string s){ return (int)s.size(); };
 
 
   BPlusTree<string> * tree = new BPlusTree<string>(keyConverter, 4);
